requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.0
outcome==1.3.0.post0
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
Telethon==1.42.0
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.5
wsproto==1.3.2
yarl==1.22.0